+++ XrmPluginCore.SourceGenerator/CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.0.1 - 27 November 2025
+* Fix: Analyzer for XPC4005 to correctly identify lambda expressions in AddImage calls
+
 ### v1.0.0 - 27 November 2025
 Initial release of XrmPluginCore SourceGenerator
 * Add: Type-Safe Images feature with compile-time enforcement via source generator