+++ XrmPluginCore/CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.2.1 - 27 November 2025
+* Fix: Analyzer for XPC4005 to correctly identify lambda expressions in AddImage calls
+
 ### v1.2.0 - 27 November 2025
 * Add: Type-Safe Images feature with compile-time enforcement via source generator
 * Add: Source analyzer rules with hotfixes and documentation to help use the Type-Safe Images feature correctly